+++ public/CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [1.4.16] - 2016-12-30
 ### Added
 - `[Backend]` getTitleID now has a $create param (defaults to TRUE)
+- `[Backend]` get_content can now be sent as POST (with fields).
 
 ## [1.4.15] - 2016-12-29
 ### Fixed